requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
anchorpy==0.21.0
anchorpy-core==0.2.0
anyio==4.9.0
attrs==25.3.0
backports.tarfile==1.2.0
base58==2.1.1
based58==0.1.1
borsh-construct==0.1.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
charset-normalizer==3.4.2
cleo==2.1.0
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
crashtest==0.4.1
distlib==0.3.9
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jito_py_rpc==0.1.4
jsonalias==0.1.1
jupiter-python-sdk==0.0.2.0
keyring==25.6.0
mnemonic==0.21
more-itertools==10.7.0
msgpack==1.1.0
multidict==6.4.4
packaging==25.0
pbs-installer==2025.5.17
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
propcache==0.3.1
pyheck==0.1.5
pyproject_hooks==1.2.0
python-dotenv==1.1.0
python-telegram-bot==22.1
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
sumtypes==0.1a6
toml==0.10.2
tomlkit==0.13.2
toolz==0.11.2
trove-classifiers==2025.5.9.12
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
websockets==15.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
